requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
transformers==4.36.0
torch>=2.2.0
accelerate==0.24.1
sentencepiece
protobuf==4.25.1
numpy==1.24.3
langgraph
